fix: group composer type ref with its implementation
